+++ src/SimpleLabeledDataset.cpp
@@ -15,7 +15,7 @@ SimpleLabeledDataset::SimpleLabeledDataset(int inputDimension, int outputDimensi
     inDimension = inputDimension;
     outDimension = outputDimension;
     this->size = size;
-    data = new float[(inputDimension + outputDimension) * size];
+    data = new double[(inputDimension + outputDimension) * size];
 }
 
 SimpleLabeledDataset::SimpleLabeledDataset(const SimpleLabeledDataset& orig) {
@@ -25,16 +25,16 @@ SimpleLabeledDataset::~SimpleLabeledDataset() {
     delete data;
 }
 
-void SimpleLabeledDataset::addPattern(const float *input, const float *output) {
+void SimpleLabeledDataset::addPattern(const double *input, const double *output) {
     
     if (addedCounter >= size) {
         LOG()->error("Trying to add %d learning patterns while the dataset size is only %d.", addedCounter+1, size);
     }
     
-    int inputSize = sizeof(float) * inDimension;
-    int outputSize = sizeof(float) * outDimension;
+    int inputSize = sizeof(double) * inDimension;
+    int outputSize = sizeof(double) * outDimension;
     int patternSize = inDimension + outDimension;
-    float *dataPtr = data + (addedCounter * patternSize);
+    double *dataPtr = data + (addedCounter * patternSize);
     std::memcpy(dataPtr, input, inputSize);
     std::memcpy(dataPtr + inDimension, output, outputSize);
     addedCounter++;
@@ -49,10 +49,10 @@ int SimpleLabeledDataset::getOutputDimension() {
 }
 
 void SimpleLabeledDataset::initDataset() {
-    data = new float[size * (inDimension + outDimension)];
+    data = new double[size * (inDimension + outDimension)];
 }
 
-float* SimpleLabeledDataset::next() {
+double* SimpleLabeledDataset::next() {
     return data + (cursor++ * (inDimension + outDimension));
 }
